+++ index/miner/lotusidx/miner_test.go
@@ -1,4 +1,4 @@
-package module
+package lotusidx
 
 import (
 	"context"
